% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conf.R
\name{adjust.q}
\alias{adjust.q}
\title{A simple function to apply the qvalue correction to a set of pvalues}
\usage{
adjust.q(p, fdr, lambda)
}
\arguments{
\item{p}{a vector of p.values to be passed to qvalue::qvalue()}

\item{fdr}{the false discovery rate}

\item{lambda}{The value of the tuning parameter to estimate \eqn{pi_0}. Must be in \eqn{[0,1)}}
}
\value{
an \eqn{n X 2} dataframe containing the qvalues, and significance (logical)
}
\description{
A simple function to apply the qvalue correction to a set of pvalues
}
