% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.vec.R
\name{class.vec}
\alias{class.vec}
\title{A function to classify the indicator vector output by infer.trio()}
\usage{
class.vec(vec = NULL)
}
\arguments{
\item{vec}{a binary vector of minimum length 6}
}
\description{
This function takes in the output from infer.trio() and classifies it as one of the 5 model topologies given by
Badsha and Fu, 2019. If the no topology is matched, it returns "Other"
}
