% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CNAtrios}
\alias{CNAtrios}
\title{100 CNA trios where the first column of each trio corresponds to the Copy Number Alteration,
the second and third columns are the expression and methylation targets of the genetic variant. The remaining
columns correspond to each trios identified confounding variables}
\format{
a list of length = 100 where each element is an 564 X 3+n dataframe where the first
column is the CNA representing V1 and columns 2:3 are the expression and Methylation targets representing
T1 and T2 (respectively) and the remaining columns are selected confounding covariates
}
\usage{
CNAtrios
}
\description{
A 670 X 670 dataframe of PCA scores where the rows are observations and the columns are the PCs

A list of dataframes where each dataframe represents a single CNA trio with confounders
}
\references{
The example data here are taken in whole or part from data generated by the TCGA Research Network: https://www.cancer.gov/tcga."
}
\keyword{datasets}
