% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.cats.R
\name{convert.cats}
\alias{convert.cats}
\title{A function to simplify inferred model classes}
\usage{
convert.cats(input = NULL)
}
\arguments{
\item{input}{a vector of inferred classes as returned by class.vec()}
}
\description{
This function takes in a vector of inferred classes from class.vec() and converts them to one of M0, M1, M2, M3, M4
}
