% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.truth.R
\name{convert.truth}
\alias{convert.truth}
\title{A function to convert the value of 'model' from MRPC::simulateData()}
\usage{
convert.truth(input = NULL)
}
\arguments{
\item{input}{a vector of true model classes}
}
\description{
This function takes in a vector of true model classes of the type 'model0', 'model1','model2'... as is input to simulateData() in the MRPC package
and returns one of "M0","M1","M2"....
}
