% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleTrios}
\alias{exampleTrios}
\title{10,000 simulated trios from the MRPC package simulateData()}
\format{
a list of length = 10,000 and each element is an n X 3 dataframe where V1 is a simulated eQTL and T1 and T2 are simulated molecular phenotypes
}
\usage{
exampleTrios
}
\description{
A list of dataframes where each dataframe represents a single trio from one of the 5 model topologies
}
\keyword{datasets}
