% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trios_with_conf.R
\name{find.parents}
\alias{find.parents}
\title{A function to find the parents of a given node in the graph}
\usage{
find.parents(Adjacency, location)
}
\arguments{
\item{Adjacency}{an adjacency matrix for the graph}

\item{location}{the column index of the node of interest in the adjacency matrix}
}
\value{
a list of length 6 representing the 6 nodes types, V,T,K,U,W,Z
}
\description{
This function uses the adjacency matrix of a graph to identify the parents of a desired node in the graph
}
