% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trios_with_conf.R
\name{gen.graph.skel}
\alias{gen.graph.skel}
\title{A function to simulate a graph with confounder, intermediate, and common child variables}
\usage{
gen.graph.skel(
  model,
  b.snp,
  b.med,
  conf.num.vec,
  number.of.T,
  number.of.V,
  struct,
  degree,
  method = "er",
  plot.graph = TRUE,
  arrow.size = 0.2,
  layout.graph = "nicely",
  neg.freq = 0.5,
  conf.coef.ranges = list(K = c(0.01, 0.1), U = c(0.15, 0.5), W = c(0.15, 0.5), Z = c(1,
    1.5))
)
}
\arguments{
\item{model}{a string specifying one of "model0","model1", "model2", "model3","model4", or "custom"}

\item{b.snp}{a numeric or vector giving the effect(s) of the genetic variant(s).}

\item{b.med}{a numeric or vector giving the effect(s) between the molecular phenotypes(s).}

\item{conf.num.vec}{a numeric vector of length 4 containing the number of known confounders (K),
unknown confounders (U), intermediate (W), and common child variables (Z; in that order).
To exclude a variable type input a zero at the given position. Note: intermediate variables (W) cannot
be included for model4 and must excluded (set to zero)}

\item{number.of.T}{a numeric indicating the number of T variables desired for model == "custom" only}

\item{number.of.V}{a numeric indicating the number of V variables desired for model == "custom" only}

\item{struct}{For use when when model == "custom". Either (1) a sub-adjacency matrix of dimension (number.of.V + number.of.T X number.of.V + number.of.T) definining the topology of the V and T nodes or (2) the string "random" denoting a random topology}

\item{degree}{passed to get.custom.graph()}

\item{method}{passed to get.custom.graph()}

\item{plot.graph}{(logical) if TRUE the graph is plotted. default = TRUE}

\item{arrow.size}{a numeric value giving the desired size of the edge arrows when plot.graph == TRUE}

\item{layout.graph}{a string specifying one of 'nicely' or 'as tree'. Determines the plotting structure of the graph
when plot.graph == TRUE. default = 'nicely'. Note that 'as tree' may be better when the graph is greater than 30 nodes.}

\item{neg.freq}{the frequency of negative effects for simulated effects. passed to gen.conf.coefs()}

\item{conf.coef.ranges}{a list of length 4 representing the U,K,W, and Z (in that order) node effects where each list
element is a vector of length 2 giving the minimum and maximum effect sizes for the given node. default values follow
from the simulations of Yang et. al., 2017. passed to gen.conf.coefs()}
}
\value{
a list of length 3
\describe{

\item{adjacency}{the indicator adjacency matrix of the graph}
\item{effects.Adj}{the adjacency matrix with the effects in place of the indicator values}
\item{igraph.obj}{an igraph object}
\item{true.adj}{only returned when model = "model4". It is the true adjacency matrix of the graph (the graph plotted plot.graph = TRUE) where as the
the graph in $adjacency is the graph used to emulate the model 4 network}

}
}
\description{
This function simulates a graph adjacency matrix for a desired topology such as M0 - M4, custom, or random topological
graphs such that they include known confounder (denoted K), unkown confounder (denoted U), intermediate (denoted W), and common child
variables (denoted Z). This function wraps gen.conf.coefs() and get.custom.graph()
}
\examples{
# generate a model 1 graph with one of each type of confounding variables
adj = gen.graph.skel(model = "model1",
                     b.snp=0.8,
                     b.med=0.6,
                     plot.graph = FALSE,
                     conf.num.vec = c(K = 1, U = 1, W = 1, Z = 1))$adjacency

# now plot the graph
plot.graph.from.adj(adj, return.igraph = FALSE)




# generate a custom/random directed graph with 2 variants and 4 molecular phenotypes
# and one of each type of confounding variable
\dontrun{
X=gen.graph.skel(model = "custom",
                 b.snp = c(0.5,0.6,0.7),
                 b.med = c(0.2,0.1,0.6),
                 conf.num.vec = c(K = 1, U = 1, W = 1, Z = 1),
                 number.of.T = 4,
                 number.of.V = 2,
                 struct = "random",
                 degree = 3,
                 method = "er",
                 plot.graph = TRUE,
                 neg.freq =0.5,
                 conf.coef.ranges=list(K=c(0.01, 0.1),
                                       U=c(0.15,0.5),
                                       W=c(0.15,0.5),
                                       Z=c(1, 1.5)))
}




}
