% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.adj.R
\name{get.adj}
\alias{get.adj}
\title{A function to get the adjacency matrix from the edge inference returned by infer.trio()}
\usage{
get.adj(inf)
}
\arguments{
\item{inf}{a vector of length 14 as return by infer.trio()}
}
\value{
a 3 X 3 adjacency matrix for the trio
}
\description{
This function takes in the output from infer.trio() and converts it to an adjacency matrix for the trio
}
\examples{
\dontrun{
inf = infer.trio(M1trio)
Adj = get.adj(inf)
}
}
