% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conf.R
\name{get.conf.matrix}
\alias{get.conf.matrix}
\title{A function to select a set of confounding variables for an input dataset}
\usage{
get.conf.matrix(
  data = NULL,
  cov.pool = NULL,
  blocksize = 2000,
  apply.qval = TRUE,
  selection_fdr = 0.05,
  lambda = NULL,
  alpha = 0.05,
  save.list = FALSE,
  save.path = "/path/to/save/location"
)
}
\arguments{
\item{data}{a matrix or dataframe of size \eqn{n samples X p variables}.}

\item{cov.pool}{a matrix or dataframe of size \eqn{n samples X q covariates}whose columns represent the possible confounding variables to be selected from}

\item{blocksize}{the number of columns of \eqn{data} to use in each block of correlation calculations passed to propagate::bigcor}

\item{apply.qval}{(logical) \eqn{default = TRUE} applies the qvalue adjustment to each set of correlations between a PC
and the columns of \eqn{trios}. If \eqn{FALSE}, the Bonferroni correction is applied.}

\item{selection_fdr}{the false discovery rate (default = 0.05) for selecting confounders when apply.qval = TRUE}

\item{lambda}{When apply.qval = TRUE, lambda is the set of cut off points of the tuning parameter to estimate \eqn{pi_0}. Must be between 0,1. If is.null(lambda) the default
passed to \eqn{adjust.q()} is \eqn{seq(0.5, max(pvalues), 0.05)}}

\item{alpha}{the test threshold for the bonferroni correacted pvalues when \eqn{apply.qval = FALSE}}

\item{save.list}{(logical) if TRUE the output is saved as a .RData object (default = FALSE)}

\item{save.path}{string specifying the path name of the output list to be save as a .RData structure}
}
\value{
a list of 6 elements containing:
\describe{
\item{sig.asso.covs}{a list of \eqn{length=ncol(data)}. Each element in the list is a vector of of indices corresponding to the columns
of \eqn{cov.pool} that are significantly associated to each input variable (column) in \eqn{data}}
\item{pvalues}{a matrix of dimension \eqn{ncol(cov.pool) X ncol(data)} of the pvalues from the pearson correlation test on each set of correlations}
\item{qvalues}{a matrix of dimension \eqn{ncol(cov.pool) X ncol(data)} of qvalues from each set of correlations}
\item{cors}{the matrix of the calculated pairwise correlations of dimension \eqn{ncol(cov.pool) X ncol(trios)}}
\item{sig}{A matrix of logical values of dimension \eqn{ncol(cov.pool) X ncol(data)}}
\item{adj.p}{A matrix of dimension \eqn{ncol(cov.pool) X ncol(data)} of the adjusted p-values if \eqn{apply.qval = FALSE}}
}
}
\description{
This function takes in a set of variables represented in the columns of \eqn{data} and a pool of potential confounding variables (e.g whole-genome expression or
methylation PC scores) and selects the confounding variables associated with each column \eqn{data} by testing the marginal associations of each column of \eqn{data} with
the covariates in \eqn{cov.pool}. This method is a simpler version of \eqn{get.conf.trios} which does not require the input data be
organized into trios.
}
\examples{

\dontrun{
#fast example on a set of genes
#set the blocksize to be 1/3 the number of genes
gene.confs=get.conf.matrix(data=WBgenes,
                          cov.pool=WBscores,
                          blocksize=round((1/3)*dim(WBgenes)[2]))

#fast example on a set of SNPs
#set the blocksize to be 1/3 the number of snps
snp.confs=get.conf.matrix(data=WBsnps,
                          cov.pool=WBscores,
                          blocksize=round((1/3)*dim(WBsnps)[2]))

}
}
\references{
\preformatted{\\insertAllCited\{\}
}
}
