% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conf.R
\name{get.conf}
\alias{get.conf}
\title{A function to obtain confounders for trios using the PCs from the whole-genome expression or methylation data}
\usage{
get.conf(
  trios = NULL,
  PCscores = NULL,
  FDR = 0.1,
  alpha = 0.05,
  save.list = FALSE,
  return.list = TRUE,
  save.path = "/path/to/save/location"
)
}
\arguments{
\item{trios}{either (1) a list where each element contains a dataframe corresponding to a trio or (2) a single data
frame with samples in rows and trios in the columns}

\item{PCscores}{a dataframe of the whole-genome PCA scores with observations in rows and PCs in columns}

\item{FDR}{the false discovery rate (default = 0.10)}

\item{alpha}{the alpha level of the pearson correlation test used by psych::corr.test() (default = 0.05)}

\item{save.list}{(logical) if TRUE the output is saved as a .RData object (default = FALSE)}

\item{return.list}{(logical) if TRUE the list of the column indices of significant PCs detected for each trio
is returned (default = TRUE)}

\item{save.path}{string specifying the path name of the output}
}
\value{
a list of the column indices of significant PCs detected for each trio
}
\description{
This function takes in a list of trios and the scores from the PCA on whole-genome level expression
or methylation data and performs PC selection
}
\examples{
#fast example on 40 trios
trio.with.conf=get.conf(trios=WBtrios[1:40], PCscores=WBscores)

}
