% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conf.R
\name{get.conf.trios}
\alias{get.conf.trios}
\title{A function to select a set of confounding variables for each trio}
\usage{
get.conf.trios(
  trios = NULL,
  cov.pool = NULL,
  blocksize = 2000,
  selection_fdr = 0.05,
  filter_int_child = FALSE,
  filter_fdr = 0.1,
  adjust_by = c("individual", "all", "none"),
  alpha = 0.01,
  lambda = NULL,
  pi0.method = "smoother",
  save.list = FALSE,
  save.path = "/path/to/save/location"
)
}
\arguments{
\item{trios}{either (1) a list where each element contains a dataframe corresponding to a trio or (2) a single data
frame with samples in rows and trios in the columns. The assumed structure of each trio is that the Genetic variant
is the first column followed by the two molecular phenotypes.}

\item{cov.pool}{a dataframe of candidate confounding variables with observations in rows and covariates in columns}

\item{blocksize}{the number of variants to use in each block of correlation calculations passed to propagate::bigcor}

\item{selection_fdr}{the false discovery rate (default = 0.05) for selecting confounders}

\item{filter_int_child}{(logical) to determine if common child and intermediate confounding variables should be filtered out
(removed) from the significant confounders for each trio. Note: only used when return.for.trios==TRUE. Default = FALSE.}

\item{filter_fdr}{the false discovery rate for filtering common child
and intermediate confounding variable.}

\item{adjust_by}{a string specifying one of 'individual', 'all', or 'none'. If 'individual' then the multiple comparisons adjustment is done for the set of tests corresponding
to all covariates in \eqn{cov.pool} with each column of \eqn{data}. If 'all' then the adjustment is done using all pvalues.
if 'none', the pvalues for the selecting confounders are not adjusted and taken at the threshold alpha.}

\item{alpha}{the pvalue cutoff when adjust_by = 'none'. default is alpha = 0.01}

\item{lambda}{the cut off points of the tuning parameter to estimate \eqn{pi_0}. Must be between 0,1. If is.null(lambda) the default
passed to \eqn{adjust.q()} is \eqn{seq(0.5, max(pvalues), 0.05)}}

\item{pi0.method}{a string specifying one of 'smoother' or 'boostrap' describing the method used to estimate
Pi0. Passed to qvalue::qvalue(). default = 'smoother'. Note: bootstrap' can be used in place of 'smoother' if 'smoother' fails}

\item{save.list}{(logical) if TRUE the output is saved as a .RData object (default = FALSE)}

\item{save.path}{string specifying the path name of the output}
}
\value{
a list of 9 elements containing:
\describe{
\item{sig.asso.covs}{a list of length = number of trios. each element is a set of column indices indicating selected covariates
from \eqn{cov.pool}}
\item{filtered.covs}{if \eqn{filter_in_child} = TRUE, then it is a list of length = number of trios giving the column indices of
\eqn{cov.pool} indicating the covariates that were filtered}
\item{reg.pvalues}{a matrix of dimension \eqn{number of trios X ncol(cov.pool)} of pvalues}
\item{reg.qvalues}{a matrix of dimension \eqn{number of trios X ncol(cov.pool)} of qvalues}
\item{reg.sig.mat}{a logical matrix of dimension \eqn{number of trios X ncol(cov.pool)} indicating the significant qvalues from \eqn{reg.q.values}}
\item{filtering.pvalues}{a matrix of dimension \eqn{number of trios X ncol(cov.pool)} of filtering pvalues}
\item{filtering.correlations}{a correlation matrix of dimension \eqn{number of trios X ncol(cov.pool)} representing the correlations between the genetic variants and covariates}
\item{filtering.qvalues}{a matrix of dimension \eqn{number of trios X ncol(cov.pool)} of filtering qvalues}
\item{filtering.sigmat}{a logical matrix of dimension \eqn{number of trios X ncol(cov.pool)} indicating the significant qvalues from \eqn{filtering.qvalues}}
}
}
\description{
This function takes in a list or dataframe of trios and a matrix of potential confounding variables (e.g whole-genome expression or
methylation PC scores) and selects the confounding variables associated with each trio using the regression-based procedure of
\insertCite{yang2017identifying}{MRGN}. Additionally, this function allows the user to filter common child and intermediate confounding
variables by removing selected confounders that are significantly associated with the genetic variant.
}
\examples{

\dontrun{
#example 1 (no filtering, no FDR correction)
trio.conf=get.conf.trios(trios=WBtrios,
                         cov.pool=WBscores,
                         blocksize=20,
                         adjust_by = 'none')

#example 2 (filtering of common child/intermediate variables)
trio.conf2=get.conf.trios(trios=WBtrios,
                         cov.pool=WBscores,
                         blocksize=20,
                         filter_int_child = TRUE,
                         adjust_by = 'all')
}
}
\references{
\insertRef{yang2017identifying}{MRGN}
}
