% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.metrics.R
\name{get.metrics}
\alias{get.metrics}
\title{A function to compute the precision and recall between two graphs}
\usage{
get.metrics(
  Truth,
  Inferred,
  reg.vec,
  nV = 1,
  includeV = TRUE,
  weight.edge.directed.present = 1,
  weight.edge.present.only = 0.5,
  get.adj.inf = FALSE,
  get.adj.truth = FALSE
)
}
\arguments{
\item{Truth}{the true model class or adjacency matrix}

\item{Inferred}{the inferred model class or adjacency matrix}

\item{reg.vec}{the vector of regression indicators passed to get.adj.from.class() when the inferred input
is a class label (i.e. when get.adj.inf = TRUE)}

\item{nV}{the number of genetic variants in the graph (default = 1 for trios, passed to MRPC::RecallPrecision)}

\item{includeV}{(logical) whether to include the edges with the genetic variant V (default = TRUE, passed to MRPC::RecallPrecision)}

\item{weight.edge.directed.present}{a numeric specifying the weight applied to true positive edges with correct directions
default value is 1 (passed to MRPC::RecallPrecision)}

\item{weight.edge.present.only}{a numeric specifying the weight applied to true positive edges with incorrect directions.
default is 0.5 (passed to MRPC::RecallPrecision)}

\item{get.adj.inf}{(logical) use TRUE if inferred inputs are class labels. default = FALSE}

\item{get.adj.truth}{(logical) use TRUE if truth inputs are class labels. default = FALSE}
}
\value{
a named vector
\describe{
\item{Precision}{The edge-based precision.}
\item{Recall}{The edge-based recall.}
}
}
\description{
This function takes as input true and inferred model classes as is returned by class.vec() or their
adjacency matrices. This function wraps MRGN::get.adj.from.class() and MRPC::RecallPrecision()
}
\examples{
\dontrun{
#example using model class labels as input:
result = get.metrics(Truth = "M4", Inferred = "M3", get.adj.inf = TRUE, get.adj.truth = TRUE)

print(result)

#example using adjacency matrix as input
true = get.adj.from.class('M1')
pred = get.adj.from.class('M4')

result = get.metrics(Truth = true, Inferred = pred)

print(result)
}
}
\references{
\insertRef{badsha2021mrpc}{MRGN}
}
