% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conf.R
\name{get.q.sig}
\alias{get.q.sig}
\title{this function wraps adjust.q and adjust.q reg}
\usage{
get.q.sig(
  pvalues,
  fdr.level,
  lambda.seq,
  pi0.method,
  adjustment,
  contains.na = FALSE
)
}
\arguments{
\item{pvalues}{a matrix of pvalues or a list of 2-list containing the pvalues and indicies for which pvalues to exlude
(i.e from filtering)}

\item{fdr.level}{the false discovery rate passed to adjust.q or adjust.q.reg}

\item{lambda.seq}{The value(s) of the tuning parameter to estimate \eqn{pi_0}. Must be in \eqn{[0,1]} passed to adjust.q or adjust.q.reg}

\item{pi0.method}{A string specifying 'smoother' or 'bootstrap'. Passed to qvalue::qvalue()}

\item{adjustment}{a string specifying one of 'individual' (apply the qvalue correction separately to each
column of pvalues) or 'all' (apply the correction jointly to all pvalues)}

\item{contains.na}{(logical) TRUE when pvalues contains NA values from filtered covariates, in which case
adjust.q.reg is used in place of adjust.q. default = FALSE}
}
\value{
of input.type = list, then the output is a list of 2-lists containing the q values and sigificance. Else if the
input.type == array, then the output is a 2-list where the first element is a matrix of qvalues and the second element is
a matrix of significance determinations
}
\description{
It is not recommended to use this function directly
}
