% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRData.R
\name{loadRData}
\alias{loadRData}
\title{A simple helper function to read in .Rdata files to a specific handle}
\usage{
loadRData(fileName = NULL)
}
\arguments{
\item{fileName}{path/to/.Rdata}
}
\description{
This function utilizes the load() function to read in a .Rdata or .RData file with a specific handle
}
