% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{M1trio}
\alias{M1trio}
\title{An M1 trio with 1000 observations}
\format{
dataframe 1000 X 3
}
\usage{
M1trio
}
\description{
A 1000 X 3 dataframe where V1 is a simulated eQTL and T1 and T2 are simulated molecular phenotypes
}
\keyword{datasets}
