% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGN-package.R
\docType{package}
\name{MRGN-package}
\alias{MRGN}
\alias{MRGN-package}
\title{MRGN: Medenlian Randomization Genomic Network}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package implements the MRGN (Mendelian Randomization for Genomic Networks) method. This method leverages the principle of Mendelian randomization to infer causal relationship for a genomic trio, which consists of a genetic variant (as the instrumental variable) and two molecular phenotypes (for example, expression of two genes), with or without the presence of confounders.
}
\author{
\strong{Maintainer}: Jarred Kvamme \email{jarredk@uidaho.edu} (\href{https://orcid.org/0000-0001-7333-2370}{ORCID})

Authors:
\itemize{
  \item Audrey Fu \email{audreyqyfu@gmail.com}
}

}
\keyword{internal}
