% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conf.R
\name{p.from.parcor}
\alias{p.from.parcor}
\title{A function to calculate the partial correlation test pvalues using fishers z-transformation}
\usage{
p.from.parcor(r, n, S)
}
\arguments{
\item{r}{either (1) a single partial correlation coefficient or (2) a vector of partial correlation coefficients}

\item{n}{the sample size (if \eqn{length(r)>1} then n must be a vector of sample sizes)}

\item{S}{the number of variables in the conditioning set}
}
\value{
an \eqn{n X 2} dataframe containing the partial correlations and pvalues
}
\description{
A function to calculate the partial correlation test pvalues using fishers z-transformation
}
