% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conf.R
\name{p.from.reg}
\alias{p.from.reg}
\title{A function to perform the regression of a possible confounding pc aganst all pairs of cis and trans genes given
in \eqn{genes}}
\usage{
p.from.reg(inlist, cov.cols)
}
\arguments{
\item{inlist}{a list of 3 elements}

\item{cov.cols}{the column dimension of the covariate pool}
}
\value{
a vector
}
\description{
A function to perform the regression of a possible confounding pc aganst all pairs of cis and trans genes given
in \eqn{genes}
}
