% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trios_with_conf.R
\name{plot.graph.from.adj}
\alias{plot.graph.from.adj}
\title{A function to visualize the DAG from a simulated adjacency matrix. This function uses igraph to convert
The adjacency matrix to}
\usage{
\method{plot}{graph.from.adj}(
  Adj,
  return.igraph = TRUE,
  graph.layout = c("nicely", "as tree"),
  node.size = NULL,
  arrow.size = NULL,
  node.color = NULL,
  ...
)
}
\arguments{
\item{Adj}{an n X n adjacency matrix}

\item{return.igraph}{(logical) if TRUE the igraph object is returned. default == TRUE}

\item{graph.layout}{a string specifying one of 'nicely' or 'as tree'. Determines the plotting structure of the graph}

\item{node.size}{an integer giving the desired plotting size for graph vertices. Passed to igraph::plot_igraph}

\item{arrow.size}{an integer giving the desired size for direct edges. Passed to igraph::plot_igraph}

\item{node.color}{either a string giving the desired color or a vector of length n giving the colors for each
node. Passed to igraph::plot_igraph}

\item{...}{additional plotting parameters. See igraph::plot.igraph}
}
\value{
igraph object
}
\description{
A function to visualize the DAG from a simulated adjacency matrix. This function uses igraph to convert
The adjacency matrix to
}
