% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conf.R
\name{q.from.cor}
\alias{q.from.cor}
\title{A function to calculate the pearson correlation p-values and apply the qvalue correction}
\usage{
q.from.cor(r, n, fdr)
}
\arguments{
\item{r}{either (1) a single pearson correlation coefficient or (2) a vector of correlation coefficients}

\item{n}{the sample size (if length(r)>1 then n must be a vector if sample sizes vary)}

\item{fdr}{the false discovery rate}
}
\value{
a logical vector indicating significant correlations after qvalue correction
}
\description{
A function to calculate the pearson correlation p-values and apply the qvalue correction
}
