% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg.R
\name{Reg}
\alias{Reg}
\title{A function which preforms the standard regressions between the nodes in the trio. Not intended for direct use}
\usage{
Reg(data = NULL, verbose = FALSE)
}
\arguments{
\item{data}{A dataframe with at least 3 columns and the first column containing the genetic variant}

\item{verbose}{(logical) if TRUE the summary of the regressions is printed

PermReg()}
}
\description{
A function which preforms the standard regressions between the nodes in the trio. Not intended for direct use
}
