% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trios_with_conf.R
\name{simData}
\alias{simData}
\title{A function to simulate a trio with confounding variable effects}
\usage{
simData(theta, model, b0.1, b.snp, b.med, sd.1, G = NULL, u = NULL)
}
\arguments{
\item{theta}{the frequency of the alternative allele where the reference is (1-theta)}

\item{model}{a string specifying one of "model0","model1", "model2", "model3","model4"}

\item{b0.1}{the intercept}

\item{b.snp}{the effect of the genetic variant}

\item{b.med}{the effect between the molecular phenotypes}

\item{sd.1}{the residual standard deviation (noise)}

\item{G}{A matrix of size N X g of confounding variables}

\item{u}{either a numeric specifying the number of confounders to be selected at random from G or
the indices of the confounders to be selected.}
}
\value{
a matrix of dimension N X (3+g) representing the simulated trio and confounding variables
}
\description{
This function simulates a trio from one of the model toplogies given by Badsha and Fu, 2019 with counding effects on
the molecular phenotypes. Note the effects of the confounding variables are selected from uniform(0.15,0.5) Yang et al. 2017
}
