% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trios_with_conf.R
\name{simDataNP}
\alias{simDataNP}
\title{A function to simulate a child node with no parents and confounding variables}
\usage{
simDataNP(N, b0.1, sd.1, G = NULL, u = NULL, gamma.u = NULL)
}
\arguments{
\item{N}{Simulation sample size}

\item{b0.1}{the intercept}

\item{sd.1}{the residual standard deviation (noise)}

\item{G}{A matrix of size N X g of confounding variables}

\item{u}{the indices of the confounders to be selected. If NULL confounders are omitted}

\item{gamma.u}{a vector of length g specifying the effects of the confounding variables}
}
\value{
a vector of length N containing the simulated values of the child node
}
\description{
A function to simulate a child node with no parents and confounding variables
}
