% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synData}
\alias{synData}
\title{100 simulated graphs with confounding, intermediate, and common child variables}
\format{
a list of length = 100 where each element is a 4-list
\describe{
\item{data}{a 500 X (3 + n) dataframe where the first three columns are V1, T1, and T2 and the
remaining n columns are the simulated confounding variables. Column names are suffixed with the
index of the dataset.}
\item{Adjacency}{the adjacency matrix of the simulating graph.}
\item{Effects}{a matrix of the same dimension as Adjacency giving the simulated effect size of
each edge.}
\item{igraph}{the igraph representation of the simulating graph.}
}
}
\usage{
synData
}
\description{
The full output of simData.from.graph() for each of 100 simulated datasets. Each dataset
was simulated under one of the 5 model topologies with 1 to 5 confounders (U), one intermediate
variable (W), and one common child (Z). The simulation parameters are recorded in synParams.
}
\seealso{
\link{synTrios}, \link{synConfs}, \link{synParams}, \link{simData.from.graph}
}
\keyword{datasets}
