% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synParams}
\alias{synParams}
\title{The simulation parameters used to generate synData}
\format{
a 100 X 14 dataframe where each row corresponds to one simulated dataset
\describe{
\item{V1, V2, V3, V4, V5, V6}{unused placeholder columns retained from the pre-allocated
parameter matrix; all values are zero.}
\item{model}{the model topology the dataset was simulated under, one of 'model0' through 'model4'.}
\item{minor.freq}{the minor allele frequency of the simulated variant.}
\item{b.snp}{the effect of the variant on the molecular phenotypes.}
\item{b.med}{the mediation effect between the molecular phenotypes.}
\item{SD}{the standard deviation of the residual noise.}
\item{num.confounder}{the number of confounding (U) variables.}
\item{num.intermediate}{the number of intermediate (W) variables.}
\item{num.common.child}{the number of common child (Z) variables.}
}
}
\usage{
synParams
}
\description{
The randomly drawn parameters used to simulate each of the 100 datasets in synData
}
\seealso{
\link{synData}, \link{synTrios}, \link{synConfs}
}
\keyword{datasets}
