% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synTrios}
\alias{synTrios}
\title{The 100 simulated trios from synData}
\format{
a list of length = 100 where each element is a 500 X 3 dataframe containing the simulated
variant V1 and the two molecular phenotypes T1 and T2
}
\usage{
synTrios
}
\description{
The first three columns (V1, T1, T2) of each simulated dataset in synData
}
\seealso{
\link{synData}, \link{synConfs}, \link{synParams}
}
\keyword{datasets}
