% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WBscores}
\alias{WBscores}
\title{The rotated variables from the whole-genome level expression PCA for the GTEx tissue Whole Blood}
\format{
dataframe 670 X 670
}
\usage{
WBscores
}
\description{
A 670 X 670 dataframe of PCA scores where the rows are observations and the columns are the PCs
}
\keyword{datasets}
