% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WBsnps}
\alias{WBsnps}
\title{The eQTL columns of the 100 GTEx Whole Blood trios in WBtrios}
\format{
a 670 X 100 matrix where the rows are observations and each column is the eQTL of one trio
}
\usage{
WBsnps
}
\description{
The V1 (eQTL) column extracted from each of the first 100 trios in WBtrios and bound
column-wise into a single matrix
}
\references{
The GTEx Consortium. GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2). https:// gtexportal.org/home/datasets, 2021. (Online; accessed 8-June-2020)
}
\seealso{
\link{WBtrios}, \link{WBgenes}
}
\keyword{datasets}
