% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WBtrios}
\alias{WBtrios}
\title{200 trios formed from the GTEx tissue WholeBlood}
\format{
a list of length = 200 where each element is an 670 X 3 dataframe where the first
column is an eQTL representing V1 and columns 2:3 are the cis and trans targets representing
T1 and T2 (respectively)
}
\usage{
WBtrios
}
\description{
A list of dataframes where each dataframe represents a single trio from one of the 5 model topologies
}
\references{
The GTEx Consortium. GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2). https:// gtexportal.org/home/datasets, 2021. (Online; accessed 8-June-2020)
}
\keyword{datasets}
