% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conf.R
\name{adjust.q}
\alias{adjust.q}
\title{A function to calculate the pearson correlation p-values and apply the qvalue correction}
\usage{
adjust.q(p, fdr, lambda)
}
\arguments{
\item{p}{a vector of p.values to be passed to qvalue::qvalue()}

\item{fdr}{the false discovery rate}

\item{lambda}{The value of the tuning parameter to estimate \eqn{pi_0}. Must be in \eqn{[0,1)}}
}
\value{
an n X 3 dataframe containing the correlations, qvalues, and significance (logical)
}
\description{
A function to calculate the pearson correlation p-values and apply the qvalue correction
}
