% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conf.R
\name{compute.pairwise.pcors}
\alias{compute.pairwise.pcors}
\title{A function to calculate the pairwise partial correlations between variables in data and confounders in cov.pool
conditioned on the variables in cond.vars}
\usage{
compute.pairwise.pcors(data, confs, cond.vars)
}
\arguments{
\item{data}{a matrix or dataframe of variables for which to select confounders}

\item{confs}{a matrix or dataframe of covariates representing potential confounders}

\item{conf.vars}{a matrix or dataframe of variables on which to condition each pairwise relationship}
}
\value{
a matrix of size ncol(confs) x ncol(data) of partial correlations
}
\description{
A function to calculate the pairwise partial correlations between variables in data and confounders in cov.pool
conditioned on the variables in cond.vars
}
