% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conf.R
\name{filter_out}
\alias{filter_out}
\title{This function takes in the correlations with the genetic variants and performs filtering
of common child and intermediate confounding variables at the fdr threshold given by filter.fdr}
\usage{
filter_out(sig.asso.pcs, trios, PCscores, filter.fdr)
}
\arguments{
\item{sig.asso.pcs}{a matrix of correlations between the genetic variants the covariate pool of potential confounders}

\item{filter.fdr}{the filtering fdr threshold}

\item{returns}{a list of the identified common child and intermediate variables for each genetic variant in snpcors}
}
\value{
an \eqn{n X 2} dataframe containing the qvalues, and significance (logical)
}
\description{
This function takes in the correlations with the genetic variants and performs filtering
of common child and intermediate confounding variables at the fdr threshold given by filter.fdr
}
