% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trios_with_conf.R
\name{gen.conf.coefs}
\alias{gen.conf.coefs}
\title{A function to simulate effects for edges in the graph}
\usage{
gen.conf.coefs(n.effects, coef.range.list, neg.freq)
}
\arguments{
\item{n.effects}{the number of effects to simulate}

\item{coef.range.list}{a list of length 4 containing the mininum and maximum effect size for U,K,W, and Z variables
(in that order)}

\item{neg.freq}{the frequency of negative effects for a given variable}
}
\value{
a vector of simulated effects
}
\description{
This function is wrapped by gen.graph.skel() and get.custom.graph()
}
