% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trios_with_conf.R
\name{gen.graph.skel}
\alias{gen.graph.skel}
\title{A function to simulate a graph with confounder, intermediate, and common child variables}
\usage{
gen.graph.skel(
  model,
  conf.num.vec,
  number.of.T,
  number.of.V,
  struct,
  plot.graph = TRUE
)
}
\arguments{
\item{model}{a string specifying one of "model0","model1", "model2", "model3","model4", or "custom"}

\item{conf.num.vec}{a numeric vector of length 4 containing the number of confounder, known confounder, intermediate,
and common child variables (in that order). To exclude a variable type input a zero at the given position.}

\item{number.of.T}{a numeric indicating the number of T variables desired for model == "custom" only}

\item{number.of.V}{a numeric indicating the number of V variables desired for model == "custom" only}

\item{struct}{For use when when model == "custom". Either (1) a sub-adjacency matrix of dimension
(number.of.V + number.of.T  X  number.of.V + number.of.T) definining the topology of the V and T nodes
or (2) the string "random" denoting a random topology}

\item{plot.graph}{(logical) if TRUE the graph is plotted. default = TRUE}
}
\value{
an adjacency matrix
}
\description{
This function simulates a graph adjacency matrix for a desired topology such as M0 - M4, custom, or random topological
graphs such that they include confounder, intermediate, and common child variables.
}
\examples{
# generate a model 0 graph with one of each type of confounding variables
adj = gen.graph.skel(model = "model0", conf.num.vec = c(1,1,1,1))
}
