% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conf.R
\name{get.conf.matrix}
\alias{get.conf.matrix}
\title{A function to select a set of confounding variables for an input dataset}
\usage{
get.conf.matrix(
  data = NULL,
  cov.pool = NULL,
  measure = c("correlation", "partial_corr"),
  conditional.vars = NULL,
  blocksize = 2000,
  apply.qval = TRUE,
  selection_fdr = 0.05,
  adjust_by = c("individual", "all", "none"),
  lambda = NULL,
  pi0.method = "smoother",
  alpha = 0.01,
  save.list = FALSE,
  save.path = "/path/to/location"
)
}
\arguments{
\item{data}{a matrix or dataframe of size \eqn{n samples X p variables}.}

\item{cov.pool}{a matrix or dataframe of size \eqn{n samples X q covariates}whose columns represent the possible confounding variables to be selected from}

\item{measure}{a string specifying one of 'correlation' or 'partial_corr'. If 'correlation', then the marginal dependence between each column in
data and cov.pool is tested by pearson correlation. If 'partial_corr', the conditional dependence of each column of data and each column of cov.pool is
tested after conditioning on all variables in 'conditional.vars' using pearson partial correlation test. Note that 'partial_corr' is not possible if the sample size
is less than ncol(conditional.vars)+2 .}

\item{conditional.vars}{a matrix or dataframe. Used when measure = 'partial_corr'. The variables for which the user wishes to use as the conditional variable(s)
for each pairwise partial correlation estimate.}

\item{blocksize}{the number of columns of \eqn{data} to use in each block of correlation calculations passed to propagate::bigcor}

\item{apply.qval}{(logical) \eqn{default = TRUE} applies the qvalue adjustment to each set of correlations between a PC
and the columns of \eqn{trios}. If \eqn{FALSE}, the Bonferroni correction is applied. NOTE: \eqn{adjust.by } set to 'all' or 'individal'
for the adjustment to be applied}

\item{selection_fdr}{the false discovery rate (default = 0.05) for selecting confounders when apply.qval = TRUE}

\item{adjust_by}{a string specifying one of 'individual', 'all', or 'none'. If 'individual' then the multiple comparisons adjustment is done for the set of tests corresponding
to all covariates in \eqn{cov.pool} with each column of \eqn{data}. If 'all' then the adjustment is done using all pvalues.
if 'none', the pvalues for the selecting confounders are not adjusted and taken at the threshold alpha.}

\item{lambda}{When apply.qval = TRUE, lambda is the set of cut off points of the tuning parameter to estimate \eqn{pi_0}. Must be between 0,1. If is.null(lambda) the default
passed to \eqn{adjust.q()} is \eqn{seq(0.5, max(pvalues), 0.05)}}

\item{pi0.method}{a string specifying one of 'smoother' or 'boostrap' describing the method used to estimate
Pi0. Passed to qvalue::qvalue(). default = 'smoother'. NOTE: bootstrap' can be used in place of 'smoother' if 'smoother' fails}

\item{alpha}{the test threshold for adjust_by == 'none' and for the bonferroni correacted pvalues when \eqn{apply.qval = FALSE}}

\item{save.list}{(logical) if TRUE the output is saved as a .RData object (default = FALSE)}

\item{save.path}{string specifying the path name of the output list to be save as a .RData structure}
}
\value{
a list of 6 elements containing:
\describe{
\item{sig.asso.covs}{a list of \eqn{length=ncol(data)}. Each element in the list is a vector of of indices corresponding to the columns
of \eqn{cov.pool} that are significantly associated to each input variable (column) in \eqn{data}}
\item{pvalues}{a matrix of dimension \eqn{ncol(cov.pool) X ncol(data)} of the pvalues from the pearson correlation test on each set of correlations}
\item{qvalues}{a matrix of dimension \eqn{ncol(cov.pool) X ncol(data)} of qvalues from each set of correlations}
\item{cors}{the matrix of the calculated pairwise correlations (or partial correlations) of dimension \eqn{ncol(cov.pool) X ncol(trios)}}
\item{sig}{A matrix of logical values of dimension \eqn{ncol(cov.pool) X ncol(data)}}
\item{adj.p}{A matrix of dimension \eqn{ncol(cov.pool) X ncol(data)} of the adjusted p-values if \eqn{apply.qval = FALSE}}
}
}
\description{
This function takes in a set of variables represented in the columns of \eqn{data} and a pool of potential confounding variables (e.g whole-genome expression or
methylation PC scores) and selects the confounding variables associated with each column \eqn{data} by testing the marginal associations of each column of \eqn{data} with
the covariates in \eqn{cov.pool}. This method is a simpler version of \eqn{get.conf.trios} which does not require the input data be
organized into trios.
}
\examples{

\dontrun{
#fast example using 'correlation' to select confounders for Whole Blood genes
#set the blocksize to be 1/3 the number of genes
result=get.conf.matrix(data=WBgenes,
                       cov.pool=WBscores,
                       measure = 'correlation',
                       blocksize=round((1/3)*dim(WBgenes)[2]),
                       adjust_by = 'individual')




#fast example using 'partial_corr' to select PCs for Whole Blood genes conditioning on Whole Blood SNPs
result=get.conf.matrix(data=WBgenes,
                       cov.pool=WBscores,
                       conditional.vars = WBsnps,
                       measure = 'partial_corr',
                       adjust_by = 'all')

}
}
