% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conf.R
\name{get.conf}
\alias{get.conf}
\title{A function to obtain confounders for trios using the PCs from the whole-genome expression or methylation data}
\usage{
get.conf(
  trios = NULL,
  PCmat = NULL,
  alpha = 0.05,
  save.list = FALSE,
  save.path = "/path/to/save/location"
)
}
\arguments{
\item{trios}{either a list of data frames containing the trios or a single data frame with samples in rows and
trios in the columns}

\item{PCmat}{the rotated variables from the whole-genome level PCA}

\item{alpha}{the rejection threshold for the pearson correlation test between each trio and the rotated variables}

\item{save.list}{(logical) if TRUE the output is saved as a .RData object}

\item{save.path}{string specifying the path name of the output}
}
\value{
a list of dataframes where each element is a dataframe corresponding to a trio and its confounding PCs
}
\description{
This function takes in a list of trios and the rotated variables from the PCA on whole-genome level expression
or methylation data and performs PC selection
}
\examples{
#fast example on 40 trios
trio.with.conf=get.conf(trios=WBtrios[1:40], PCmat=WBscores)

#slow example for a list of trios
trios.with.confs=get.conf(trios=WBtrios, PCmat=WBscores)
}
