% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.freq.R
\name{get.freq}
\alias{get.freq}
\title{A function to calculate the minor allele frequency}
\usage{
get.freq(V = NULL)
}
\arguments{
\item{V}{a vector of alternative allele counts at a given eQTL}
}
\description{
This function takes in a vector V representing the counts of alternative alleles at an eQTL.
where 0 = homozygous for the reference allele, 1 = heterozygous for the reference allele,
and 2 = homozygous for the alternative allele. It returns the min(alternative, reference)
}
