% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.metrics.R
\name{get.metrics}
\alias{get.metrics}
\title{A function to obtain the adjacency matrix from an inferred trio class}
\usage{
get.metrics(
  Truth = NULL,
  Inferred = NULL,
  reg.vec = NULL,
  get.adj.inf = FALSE,
  get.adj.truth = FALSE
)
}
\arguments{
\item{Truth}{the true model class or adjacency matrix}

\item{Inferred}{the inferred model class or adjacency matrix)}

\item{reg.vec}{used when Inferred is from MRGN, a vector of the regression output as obtained from infer.trio()}

\item{get.adj.inf}{(logical) use TRUE if inferred inputs are class labels. default = FALSE}

\item{get.adj.truth}{(logical) use TRUE if truth inputs are class labels. default = FALSE}
}
\description{
This function takes in a the true and inferred model classes as is returned by class.vec() or their adjacency matrices
and the regression results from infer.trio() and returns the Precision, Recall, Specificity and False Omission Rate for
edge-inference
}
