% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.trio.R
\name{infer.trio}
\alias{infer.trio}
\title{A wrapper function for get.freq(), Reg(), and PermReg() to infer an input trio}
\usage{
infer.trio(
  trio = NULL,
  gamma = 0.05,
  alpha = 0.01,
  nperms = 10000,
  verbose = FALSE
)
}
\arguments{
\item{trio}{A dataframe with at least 3 columns and the first column containing the genetic variant}

\item{gamma}{the minor allele frequency threshold for which the permutation test should be used}

\item{alpha}{the rejection threshold}

\item{nperms}{the number of permuations to perform for trios with rare variants}

\item{verbose}{(logical) if TRUE results of the regressions are printed}
}
\value{
a vector containin the topology and regression results for the input trio
}
\description{
This function takes in a matrix representing a trio (with or without confounding variables) and performs the
MRGN inference. it returns a vector containing the results of the coefficient and marginal tests
}
\examples{
inference on a single trio
result=infer.trio(M1trio)
get the predicted topology
which.model=class.vec(result)
print(which.model)
infer a set of 10 trios from the built in dataset exampleTrios:
result2 = sapply(exampleTrios[sample(1:10000, 10)], infer.trio)
get the predicted topology
models = apply(result2, 2, class.vec)
print(models)
}
