% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.trio.R
\name{infer.trio}
\alias{infer.trio}
\title{A function to infer the causal network of a trio}
\usage{
infer.trio(
  trio = NULL,
  use.perm = TRUE,
  gamma = 0.05,
  is.CNA = FALSE,
  alpha = 0.01,
  nperms = 100,
  verbose = FALSE
)
}
\arguments{
\item{trio}{A dataframe with at least 3 columns. The first column should be the genetic variant, and the second
and third columns the molecular phenotypes. Data in columns 4+ are treated as confounding variables}

\item{use.perm}{(logical) if TRUE the permutation test is used for trios with minor allele freq < gamma. If FALSE
test is ignored for all trios (default = TRUE)}

\item{gamma}{The minor allele frequency threshold for which the permutation test should be used such that permutation is
performed when minor allele freq < gamma (when use.perm = TRUE).}

\item{is.CNA}{(logical) for use when use.perm=TRUE and the genetic variant is a copy number alteration so that the
threshold gamma is ignored. default = FALSE}

\item{alpha}{The rejection threshold for all wald tests (default = 0.01) which is approximately the bonferroni correction}

\item{nperms}{The number of permutations to perform for trios with rare variants (default = 10,000)}

\item{verbose}{(logical) if TRUE results of the regressions are printed}
}
\value{
a dataframe of dimension 1 x 14 with the following columns:
\describe{
\item{b11}{the indicator \link{0,1} value for the conditional test T1 ~ V | T2,U}
\item{b21}{the indicator \link{0,1} value for the conditional test T1 ~ T2 | V,U}
\item{b12}{the indicator \link{0,1} value for the conditional test T2 ~ V | T1,U}
\item{b22}{the indicator \link{0,1} value for the conditional test T2 ~ T1 | V,U}
\item{V1:T2}{the indicator \link{0,1} value for the marginal test between V1 and T2}
\item{V1:T1}{the indicator \link{0,1} value for the marginal test between V1 and T1}
\item{pb11}{the p-value for the conditional test T1 ~ V | T2,U}
\item{pb21}{the p-value for the conditional test T1 ~ T2 | V,U}
\item{pb12}{the p-value for the conditional test T2 ~ V | T1,U}
\item{pb22}{the p-value for the conditional test T2 ~ T1 | V,U}
\item{pV1:T2}{the p-value for the marginal test between V1 and T2}
\item{pV1:T1}{the p-value for the marginal test between V1 and T1}
\item{Minor.freq}{the calculated frequency of the minor allele of the genetic variant}
\item{Inferred.Model}{a string indicating the inferred model type as returned by \eqn{class.vec()}}
}
}
\description{
This function takes in a matrix representing a trio (with or without confounding variables) and performs the
MRGN inference. It wraps the functions get.freq(), Reg(), PermReg(), and class.vec() to infer an input trio.
It returns a one dimensional dataframe containing the results of the coefficient and marginal tests and the
inferred model structure
}
\examples{
#inference on a single eQTL trio
result=infer.trio(M1trio)
print(result)

\dontrun{
#fast example on 10 eQTL trios from the built in dataset WBtrios
#return just the summary stats
stats = sapply(WBtrios[1:10], function(x) infer.trio(x)$Stats)
print(stats)
#return just the inferred model topology
models = sapply(WBtrios[1:10], function(x) infer.trio(x)$Inferred.Model)
print(models)
#fast example on 10 CNA trios from the build in dataset CNAtrios
models = sapply(CNAtrios[1:10], function(x) infer.trio(x)$Inferred.Model)
print(models)
}
}
