% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.trio.R
\name{infer.trio}
\alias{infer.trio}
\title{A wrapper function for get.freq(), Reg(), and PermReg() to infer an input trio}
\usage{
infer.trio(
  trio = NULL,
  gamma = 0.05,
  alpha = 0.01,
  nperms = 10000,
  verbose = FALSE
)
}
\arguments{
\item{trio}{A dataframe with at least 3 columns and the first column containing the genetic variant}

\item{gamma}{the minor allele frequency threshold for which the permutation test should be used}

\item{alpha}{the rejection threshold}

\item{nperms}{the number of permuations to perform for trios with rare variants}

\item{verbose}{(logical) if TRUE results of the regressions are printed}
}
\description{
This function takes in a matrix representing a trio (with or without confounding variables) and performs the
MRGN inference. it returns a vector containing the results of the coefficient and marginal tests
}
