% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conf.R
\name{p.from.cor}
\alias{p.from.cor}
\title{A function to calculate the pearson correlation p-values}
\usage{
p.from.cor(r, n)
}
\arguments{
\item{r}{either (1) a single pearson correlation coefficient or (2) a vector of correlation coefficients}

\item{n}{the sample size (if \eqn{length(r)>1} then n must be a vector of sample sizes)}
}
\value{
an \eqn{n X 2} dataframe containing the correlations and pvalues
}
\description{
A function to calculate the pearson correlation p-values
}
