% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermReg.R
\name{PermReg}
\alias{PermReg}
\title{this function wraps PermReg.helper.fn() to preform the permuted regression. Not intended for direct use}
\usage{
PermReg(
  trio = NULL,
  t.obs21 = NULL,
  t.obs22 = NULL,
  p11 = NULL,
  p12 = NULL,
  m = NULL
)
}
\arguments{
\item{trio}{A dataframe with at least 3 columns and the first column containing the genetic variant}

\item{t.obs21}{the t-statistic from Reg() represnting the test on beta21}

\item{t.obs22}{the t-statistic from Reg() representing the test on beta22}

\item{p11}{the p-value from the test on beta21}

\item{p12}{the p-value from the test on beta22}

\item{m}{the number of permutations to perform
PermReg()}
}
\description{
this function wraps PermReg.helper.fn() to preform the permuted regression. Not intended for direct use
}
