% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.metrics.R
\name{RecallPrecision}
\alias{RecallPrecision}
\title{This function is copied from MRPC::RecallPrecision}
\usage{
RecallPrecision(g1, g2, GV, includeGV, edge.presence = 1, edge.direction = 0.5)
}
\description{
This function is copied from MRPC::RecallPrecision
}
